backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
skyfield
apscheduler
numpy